Add AUTHORS file and setup MySQL databases
This commit adds the AUTHORS file, which lists all individuals who have contributed content to the repository. It also includes the setup_mysql_dev.sql and setup_mysql_test.sql files, which prepare MySQL servers for the project by creating databases and granting privileges to users. These changes are necessary for the proper functioning of the project.